+++ servers/api/models/index.js
@@ -1,13 +1,14 @@
-"use strict";
+'use strict';
 
 var Sequelize = require('sequelize');
-var fs = require("fs");
-var path = require("path");
+var fs = require('fs');
+var path = require('path');
+var sleep = require('sleep');
 var db = {};
 
 module.exports = db;
 
-db.init = function (nconf) {
+db.init = function (nconf, callback) {
     var sequelize = new Sequelize(
         nconf.get('postgresql:database'), // database
         nconf.get('postgresql:username'), // username
@@ -23,24 +24,27 @@ db.init = function (nconf) {
     fs
         .readdirSync(__dirname)
         .filter(function (file) {
-            return (file.indexOf(".") !== 0) && (file !== "index.js");
+            return (file.indexOf('.') !== 0) && (file !== 'index.js');
         })
         .forEach(function (file) {
-            var model = sequelize["import"](path.join(__dirname, file));
-            console.log('Creating model ' + model.name);
+            var model = sequelize.import(path.join(__dirname, file));
             db[model.name] = model;
         });
 
     Object.keys(db).forEach(function (modelName) {
-        if ("associate" in db[modelName]) {
+        if ('associate' in db[modelName]) {
             db[modelName].associate(db);
         }
     });
 
-    sequelize.sync();
-
     db.sequelize = sequelize;
     db.Sequelize = Sequelize;
 
-    console.log('Loaded models...');
+    if (nconf.get('NODE_ENV') === 'test') {
+        sequelize.sync({force: true});
+        setTimeout(callback, 100);
+    } else {
+        sequelize.sync();
+        callback();
+    }
 };